requirements.txt
Flask==1.0.2
Werkzeug<0.14.1,>=0.12.0
pillow>=5.0.0
gunicorn==19.7.1